refactor(persistence): consolidate duplicate table definitions into shared internal objects
